refactor: move Role enum to a separate file and update imports
