+++ src/routes/enums.ts
@@ -1,0 +1,6 @@
+export enum Role {
+  MAINTAINER = "MAINTAINER",
+  FACILITATOR = "FACILITATOR",
+  CONTRIBUTOR = "CONTRIBUTOR",
+  FAVORITE = "FAVORITE"
+}
